solved issue in print_dec_wz

+++ numlen.c
@@ -6,13 +6,13 @@
 /*   By: marsoare <marsoare@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/01 13:36:03 by marsoare          #+#    #+#             */
-/*   Updated: 2024/05/01 13:37:58 by marsoare         ###   ########.fr       */
+/*   Updated: 2024/05/01 14:26:55 by marsoare         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	numlen(unsigned int number)
+int	numlen(long number)
 {
 	int	len;
